testing with square waveform disturbances
This version first simulates the plant to find the drag and inertia
coefficients. Then it creates the reference model and uses its values to
calculate the PID controller coefficients. Now the controlled system is
tested with 15 square waveform force disturbances (1Hz --> 15Hz) of
amplitude 0.1N. The system is left uncontrolled for 2 s after which the
controller is turned on. The initial overshoot as well as the settling
time is calculated.